+++ src/form/v2/birth/forms/printForm/collectors.ts
@@ -15,7 +15,7 @@ import {
   FieldConfig,
   FieldType
 } from '@opencrvs/toolkit/events'
-import { InformantType } from './pages/informant'
+import { InformantType } from '../pages/informant'
 
 export const printCertificateCollectors: FieldConfig[] = [
   {
