+++ src/form/v2/birth/index.ts
@@ -25,7 +25,7 @@ import {
 import { advancedSearchBirth } from './advancedSearch'
 import { Event } from '@countryconfig/form/types/types'
 import { SCOPES } from '@opencrvs/toolkit/scopes'
-import { BIRTH_CERTIFICATE_COLLECTOR_FORM } from './forms/print-certificate'
+import { BIRTH_CERTIFICATE_COLLECTOR_FORM } from './forms/printForm/print-certificate'
 
 export const birthEvent = defineConfig({
   id: Event.V2_BIRTH,
